move pc files to metadata dir. And added metadata breadcumb
