+++ src/app/frontend/resource/metadata/priorityclass/detail/component.ts
