+++ src/app/frontend/resource/metadata/priorityclass/detail/template.html
