+++ src/app/frontend/resource/metadata/priorityclass/module.ts
